langchain/day3/requirements.txt
langchain
langchain-community
langchain-ollama
python-dotenv
faiss-cpu
pypdf
chromadb
sentence-transformers
